Changed the meaning of the uip_*input functions. They now return success when a packet is dropped; This is needed for the ENCX24J600 driver that must make a decision to return the packet or not:  It should not retain dropped packets. From Max Holtzberg

+++ net/uip/uip_internal.h
@@ -192,8 +192,8 @@ int uip_udpinput(struct uip_driver_s *dev);
 
 /* Defined in uip_udpcallback.c *********************************************/
 
-int uip_udpcallback(struct uip_driver_s *dev,
-                    struct uip_udp_conn *conn, uint16_t flags);
+uint16_t uip_udpcallback(struct uip_driver_s *dev,
+                         struct uip_udp_conn *conn, uint16_t flags);
 #endif /* CONFIG_NET_UDP */
 
 #ifdef CONFIG_NET_ICMP
